refactor: Google App Architecture 적용 패키지 구조만 변경했음
