+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
         </service>
 
         <activity
-            android:name=".ui.container.MainActivity"
+            android:name=".ui.ui.container.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
