+++ app/src/main/java/com/ping/app/PingApplication.kt
@@ -1,11 +1,11 @@
 package com.ping.app
 
 import android.app.Application
-import com.ping.app.data.repository.PingMapRepoImpl
+import com.ping.app.data.repository.map.PingMapRepoImpl
 import com.ping.app.data.repository.login.LoginRepoImpl
-import com.ping.app.domain.dao.LoginRepo
-import com.ping.app.domain.dao.PingMapRepo
-import com.ping.app.ui.util.LocationHelper
+import com.ping.app.data.repository.login.LoginRepo
+import com.ping.app.data.repository.map.PingMapRepo
+import com.ping.app.ui.ui.util.LocationHelper
 
 class PingApplication : Application() {
     override fun onCreate() {
