+++ app/src/main/java/com/ping/app/data/database/TestDatabase.kt
@@ -1,0 +1,4 @@
+package com.ping.app.data.database
+
+class TestDatabase {
+}
