+++ app/src/main/java/com/ping/app/data/database/dao/TestDao.kt
@@ -1,0 +1,4 @@
+package com.ping.app.data.database.dao
+
+class TestDao {
+}
