+++ app/src/main/java/com/ping/app/data/database/entity/TestEntity.kt
@@ -1,0 +1,4 @@
+package com.ping.app.data.database.entity
+
+class TestEntity {
+}
