--- app/src/main/java/com/ping/app/data/local/TestEntity.kt
@@ -1,4 +1,0 @@
-package com.ping.app.data.local
-
-class TestEntity {
-}
