+++ app/src/main/java/com/ping/app/data/model/Gathering.kt
@@ -1,0 +1,11 @@
+package com.ping.app.data.model
+
+data class Gathering(
+    val uid: String, // 모임 id
+    val uuid: String, // 개최자 id
+    val gatheringTime: String,
+    val title: String,
+    val content: String,
+    val longitude: Double,
+    val latitude: Double
+)
