+++ app/src/main/java/com/ping/app/data/model/User.kt
@@ -1,0 +1,5 @@
+package com.ping.app.data.model
+
+data class User(val name: String, val email:String, val region: String, val MeetingManagerUID:String){
+    constructor():this("","","","")
+}
