--- app/src/main/java/com/ping/app/data/remote/TestAPI.kt
@@ -1,4 +1,0 @@
-package com.ping.app.data.remote
-
-interface TestAPI {
-}
