+++ app/src/main/java/com/ping/app/data/repository/TestRepo.kt
@@ -1,0 +1,4 @@
+package com.ping.app.data.repository
+
+interface TestRepo {
+}
