+++ app/src/main/java/com/ping/app/data/repository/cloudfunction/CloudFunctionRepo.kt
@@ -1,0 +1,4 @@
+package com.ping.app.data.repository.cloudfunction
+
+interface CloudFunctionRepo {
+}
