+++ app/src/main/java/com/ping/app/data/repository/cloudfunction/CloudFunctionRepoImpl.kt
@@ -1,0 +1,9 @@
+package com.ping.app.data.repository.cloudfunction
+
+import com.google.firebase.functions.FirebaseFunctions
+import com.google.firebase.functions.ktx.functions
+import com.google.firebase.ktx.Firebase
+
+class CloudFunctionRepoImpl: CloudFunctionRepo {
+    private var functions: FirebaseFunctions = Firebase.functions
+}
