+++ app/src/main/java/com/ping/app/data/repository/map/PingMapRepo.kt
@@ -1,0 +1,8 @@
+package com.ping.app.data.repository.map
+
+import com.ping.app.data.model.Gathering
+
+interface PingMapRepo {
+    fun sendPingInfo(data: Gathering)
+    fun requestAddress(lat: Double, lng: Double): String
+}
