--- app/src/main/java/com/ping/app/domain/dao/PingMapRepo.kt
@@ -1,8 +1,0 @@
-package com.ping.app.domain.dao
-
-import com.ping.app.domain.model.Gathering
-
-interface PingMapRepo {
-    fun sendPingInfo(data: Gathering)
-    fun requestAddress(lat: Double, lng: Double): String
-}
