--- app/src/main/java/com/ping/app/domain/dao/TestRepo.kt
@@ -1,4 +1,0 @@
-package com.ping.app.domain.dao
-
-interface TestRepo {
-}
