--- app/src/main/java/com/ping/app/domain/model/User.kt
@@ -1,5 +1,0 @@
-package com.ping.app.domain.model
-
-data class User(val name: String, val email:String, val region: String, val MeetingManagerUID:String){
-    constructor():this("","","","")
-}
