--- app/src/main/java/com/ping/app/ui/feature/login/LoginViewModel.kt
@@ -1,9 +1,0 @@
-package com.ping.app.presentation.ui.feature.login
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class LoginViewModel:ViewModel() {
-
-}
