--- app/src/main/java/com/ping/app/ui/feature/main/MainViewModel.kt
@@ -1,6 +1,0 @@
-package com.ping.app.ui.feature.main
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel(): ViewModel() {
-}
