--- app/src/main/java/com/ping/app/ui/feature/map/PingMapViewModel.kt
@@ -1,20 +1,0 @@
-package com.ping.app.ui.feature.map
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.naver.maps.geometry.LatLng
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
-
-class PingMapViewModel : ViewModel() {
-    private val _userLocation = MutableStateFlow<LatLng?>(null)
-    val userLocation get() = _userLocation.asStateFlow()
-    
-    fun setUserLocation(currentLocation: LatLng?) {
-        viewModelScope.launch(Dispatchers.IO) {
-            _userLocation.emit(currentLocation)
-        }
-    }
-}
