+++ app/src/main/java/com/ping/app/ui/presentation/login/LoginViewModel.kt
@@ -1,0 +1,9 @@
+package com.ping.app.ui.presentation.login
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class LoginViewModel:ViewModel() {
+
+}
