+++ app/src/main/java/com/ping/app/ui/presentation/main/MainViewModel.kt
@@ -1,0 +1,6 @@
+package com.ping.app.ui.presentation.main
+
+import androidx.lifecycle.ViewModel
+
+class MainViewModel(): ViewModel() {
+}
