+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.container.MainActivity">
+        tools:context=".ui.ui.container.MainActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/main_container"
