+++ app/src/main/res/layout/fragment_login.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.feature.login.LoginFragment">
+        tools:context=".ui.ui.feature.login.LoginFragment">
 
         <!-- TODO: Update blank fragment layout -->
